requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
asyncio==3.4.3
pyrogram>=2.0.30
py-tgcalls==1.2.0b5
python-dotenv==1.0.0
tgcrypto==1.2.5
youtube-search-python
yt-dlp==2023.12.30
